refactor: HTTP 요청, 응답 클래스 패키지 이동
- 클래스 조직화를 위해 HttpRequest, HttpResponse 클래스를 message 패키지로 이동
