+++ src/main/java/codesquad/Main.java
@@ -4,9 +4,9 @@ import codesquad.error.HttpRequestParseException;
 import codesquad.error.ResourceNotFoundException;
 import codesquad.http.HttpParser;
 import codesquad.http.HttpProcessor;
-import codesquad.http.HttpRequest;
-import codesquad.http.HttpResponse;
 import codesquad.http.HttpResponseFormatter;
+import codesquad.http.message.HttpRequest;
+import codesquad.http.message.HttpResponse;
 import codesquad.socket.ClientSocket;
 import codesquad.socket.ServerSocket;
 import java.io.IOException;
