+++ src/test/java/codesquad/http/HttpResponseFormatterTest.java
@@ -2,6 +2,7 @@ package codesquad.http;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import codesquad.http.message.HttpResponse;
 import codesquad.http.property.HttpStatus;
 import codesquad.http.property.HttpVersion;
 import java.util.Collections;
